refactor: feature flags module
