+++ apps/api/src/app.module.ts
@@ -12,7 +12,6 @@ import { HealthModule } from './app/health/health.module';
 import { OrganizationModule } from './app/organization/organization.module';
 import { EnvironmentsModule } from './app/environments/environments.module';
 import { ExecutionDetailsModule } from './app/execution-details/execution-details.module';
-import { FeatureFlagsModule } from './app/feature-flags/feature-flags.module';
 import { NotificationTemplateModule } from './app/notification-template/notification-template.module';
 import { EventsModule } from './app/events/events.module';
 import { WidgetsModule } from './app/widgets/widgets.module';
@@ -36,7 +35,6 @@ const modules: Array<Type | DynamicModule | Promise<DynamicModule> | ForwardRefe
   InboundParseModule,
   OrganizationModule,
   SharedModule,
-  FeatureFlagsModule,
   UserModule,
   AuthModule,
   HealthModule,
