+++ apps/api/src/app/events/usecases/map-trigger-recipients/map-trigger-recipients.use-case.ts
@@ -20,7 +20,7 @@ import { MapTriggerRecipientsCommand } from './map-trigger-recipients.command';
 import { CreateLog, CreateLogCommand } from '../../../logs/usecases/create-log';
 import { GetTopicSubscribersCommand, GetTopicSubscribersUseCase } from '../../../topics/use-cases';
 import { InstrumentUsecase } from '@novu/application-generic';
-import { FeatureFlagCommand, GetFeatureFlag } from '../../../feature-flags/use-cases';
+import { FeatureFlagCommand, GetFeatureFlag } from '../../../shared/use-cases';
 
 interface ILogTopicSubscribersPayload {
   environmentId: EnvironmentId;
