--- apps/api/src/app/feature-flags/services/index.ts
@@ -1,2 +1,0 @@
-export { FeatureFlagsService } from './feature-flags.service';
-export { LaunchDarklyService } from './launch-darkly.service';
